chore: update README and refactor LLM client imports
- Removed the breaking change note about the default port in README.md.
- Deleted the Claude.ts file and updated the import in index.ts to reference the correct Anthropic.ts file.
